fix: rename comic-stream-cards to comic-cards to clear Stream Quota gate
The Stream Quota Impact Note check flags any changed ctf/ path containing
"stream"; here the filename's "stream" meant the chat stream, not GetStream.
The comic AI Assistant uses Postgres + polling with zero GetStream usage, so
no quota-impact note is warranted — rename the file (and its single import)
to remove the false-positive match.
